% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.nonmem2rx}
\alias{autoplot.nonmem2rx}
\title{Autoplot nonmem2rx object}
\usage{
\method{autoplot}{nonmem2rx}(
  object,
  ...,
  ncol = 3,
  nrow = 3,
  log = "",
  xlab = "Time",
  ylab = "Predictions",
  page = NULL
)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{...}{ignored parameters for \code{nonmem2rx} objects}

\item{nrow, ncol}{Number of rows and columns}

\item{log}{Should "" (neither x nor y), "x", "y", or "xy" (or "yx") be
log-scale?}

\item{xlab, ylab}{The x and y axis labels}

\item{page}{number of page(s) for the individual plots, by default
(NULL) all pages, otherwise you can list which pages you want to
print}
}
\description{
Autoplot nonmem2rx object
}
