% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCov.R
\name{nmcov}
\alias{nmcov}
\title{Read in data file}
\usage{
nmcov(file, ...)
}
\arguments{
\item{file}{path to Nonmem table file}

\item{...}{}
}
\description{
Read in data file
}
\examples{
nmcov(system.file("mods/cpt/runODE032.cov", package="nonmem2rx"))
}
\author{
Philip Delff and Matthew L. Fidler
}
