% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{nmext}
\alias{nmext}
\title{Reads the NONMEM \code{.ext} file for final parameter information}
\usage{
nmext(file)
}
\arguments{
\item{file}{File where the list is located}
}
\value{
return a list with \verb{$theta}, \verb{$eta} and \verb{$eps}
}
\description{
Reads the NONMEM \code{.ext} file for final parameter information
}
\examples{

nmext(system.file("run001.ext", package="nonmem2rx"))
}
\author{
Matthew L. Fidler
}
