% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtab.R
\name{nmtab}
\alias{nmtab}
\title{Read nonmem table file}
\usage{
nmtab(file, ...)
}
\arguments{
\item{file}{file name to read the results from}

\item{...}{other parameters passed to `data.table::fread`}
}
\value{
data frame of the read table
}
\description{
Read nonmem table file
}
\examples{
nmtab(system.file("mods/cpt/runODE032.csv", package="nonmem2rx"))
}
\author{
Philip Delff, Matthew L. Fidler
}
