% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{nmxml}
\alias{nmxml}
\title{Read a nonmem xml and create output similar to the \code{nmlst()}}
\usage{
nmxml(xml)
}
\arguments{
\item{xml}{xml file}
}
\value{
list of nonmem information
}
\description{
Read a nonmem xml and create output similar to the \code{nmlst()}
}
\examples{
nmxml(system.file("mods/cpt/runODE032.xml", package="nonmem2rx"))
}
\author{
Matthew L. Fidler
}
