% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{nmxmlCov}
\alias{nmxmlCov}
\alias{nmxmlOmega}
\alias{nmxmlSigma}
\alias{nmxmlTheta}
\title{Get the xml for debugging (without including data etc)}
\usage{
nmxmlCov(xml, xmlout, tag = "//nm:covariance")

nmxmlOmega(xml, xmlout, tag = "//nm:omega")

nmxmlSigma(xml, xmlout, tag = "//nm:sigma")

nmxmlTheta(xml, xmlout, tag = "//nm:theta")
}
\arguments{
\item{xml}{Original xml file}

\item{xmlout}{xml output (only includes xml)}
}
\value{
nothing, called for side effects
}
\description{
Get the xml for debugging (without including data etc)
}
\author{
Matthew L. Fidler
}
\keyword{internal}
