% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrec.R, R/abbrev.R, R/data.R, R/input.R,
%   R/model.R, R/omega.R, R/pro.R, R/records.R, R/sub.R, R/tab.R, R/theta.R
\name{nonmem2rxRec.abb}
\alias{nonmem2rxRec.abb}
\alias{nonmem2rxRec.pk}
\alias{nonmem2rxRec.pre}
\alias{nonmem2rxRec.des}
\alias{nonmem2rxRec.err}
\alias{nonmem2rxRec.dat}
\alias{nonmem2rxRec.inp}
\alias{nonmem2rxRec.mod}
\alias{nonmem2rxRec.ome}
\alias{nonmem2rxRec.sig}
\alias{nonmem2rxRec.pro}
\alias{nonmem2rxRec.aaa}
\alias{nonmem2rxRec}
\alias{nonmem2rxRec.default}
\alias{nonmem2rxRec.sub}
\alias{nonmem2rxRec.tab}
\alias{nonmem2rxRec.the}
\title{Record handling for nonmem records}
\usage{
\method{nonmem2rxRec}{abb}(x)

\method{nonmem2rxRec}{pk}(x)

\method{nonmem2rxRec}{pre}(x)

\method{nonmem2rxRec}{des}(x)

\method{nonmem2rxRec}{err}(x)

\method{nonmem2rxRec}{dat}(x)

\method{nonmem2rxRec}{inp}(x)

\method{nonmem2rxRec}{mod}(x)

\method{nonmem2rxRec}{ome}(x)

\method{nonmem2rxRec}{sig}(x)

\method{nonmem2rxRec}{pro}(x)

\method{nonmem2rxRec}{aaa}(x)

nonmem2rxRec(x)

\method{nonmem2rxRec}{default}(x)

\method{nonmem2rxRec}{sub}(x)

\method{nonmem2rxRec}{tab}(x)

\method{nonmem2rxRec}{the}(x)
}
\arguments{
\item{x}{Nonmem record data item, should be of class c(stdRec,
"nonmem2rx") where the stdRec is the standardized record (pro for
\verb{$PROBLEM}, etc)}
}
\value{
Nothing, called for side effects
}
\description{
Record handling for nonmem records
}
\details{
Can add record parsing and handling by creating a S3 method for each type of standardized method
}
\author{
Matthew L. Fidler
}
\keyword{internal}
