% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nminfo.R
\name{nminfo}
\alias{nminfo}
\title{Get the most accurate information you can get from NONMEM}
\usage{
nminfo(
  file,
  xml = ".xml",
  ext = ".ext",
  cov = ".cov",
  phi = ".phi",
  lst = ".lst",
  useXml = TRUE,
  useExt = TRUE,
  useCov = TRUE,
  usePhi = TRUE,
  useLst = TRUE,
  strictLst = FALSE
)
}
\arguments{
\item{file}{nonmem file, like control stream, phi.  This function
will remove the extension to try to get the right information.
It preferentially selects the most accurate estimates from the
file.}
}
\value{
list of NONMEM information
}
\description{
Get the most accurate information you can get from NONMEM
}
\examples{
nminfo(system.file("mods/cpt/runODE032.res", package="nonmem2rx"))
}
\author{
Matthew L. Fidler
}
