% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst.R
\name{nmlst}
\alias{nmlst}
\title{Reads the NONMEM list file for information}
\usage{
nmlst(file)
}
\arguments{
\item{file}{File where the list is located}
}
\value{
return a list with `$theta`, `$eta` and `$eps`
}
\description{
Reads the NONMEM list file for information
}
\examples{
nmlst(system.file("mods/DDMODEL00000322/HCQ1CMT.lst", package="nonmem2rx"))
}
\author{
Matthew L. Fidler
}
