% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem2rx.R
\name{nonmem2rx}
\alias{nonmem2rx}
\title{Convert a NONMEM source file to a rxode control}
\usage{
nonmem2rx(file)
}
\arguments{
\item{file}{NONMEM control file location}
}
\value{
rxode2 function
}
\description{
Convert a NONMEM source file to a rxode control
}
\examples{
nonmem2rx(system.file("run001.mod", package="nonmem2rx"))
}
