% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem2rx.R
\name{nonmem2rx}
\alias{nonmem2rx}
\title{Convert a NONMEM source file to a rxode model (nlmixr2-syle)}
\usage{
nonmem2rx(
  file,
  inputData = NULL,
  nonmemOutputDir = NULL,
  rename = NULL,
  tolowerLhs = TRUE,
  thetaNames = TRUE,
  etaNames = TRUE,
  cmtNames = TRUE,
  updateFinal = TRUE,
  determineError = TRUE,
  validate = TRUE,
  lst = ".lst",
  ext = ".ext"
)
}
\arguments{
\item{file}{NONMEM run file}

\item{inputData}{this is a path to the input dataset (or `NULL` to
determine from the dataset).  Often the input dataset may be
different from the place it points to in the control stream
because directories can be created to run NONMEM from a script.
Because of this, when this is specified the input data will be
assumed to be from here instead.}

\item{nonmemOutputDir}{This is a path the the nonmem output
directory.  When not `NULL` it will assume that the diretory for
the output files is located here instead of where the control
stream currently exists.}

\item{rename}{When not `NULL` this should be a named character
vector that contains the parameters that should be renamed.  For
example, if the model uses the variable `YTYPE` and has `CMT` it
isn't compatible with `rxode2`/`nlmixr2`. You can change this for
the input dataset and the model to create a new model that still
reproduces the NONMEM output by specifying
`rename=c(dvid="YTYPE")`}

\item{tolowerLhs}{Boolean to change the lhs to lower case (default:
`TRUE`)}

\item{thetaNames}{this could be a boolean indicating that the theta
names should be changed to the comment-labeled names (default:
`TRUE`). This could also be a character vector of the theta names
(in order) to be replaced.}

\item{etaNames}{this could be a boolean indicating that the eta
names should be changed to the comment-labeled names (default:
`TRUE`). This could also be a character vector of the theta names
(in order) to be replaced.}

\item{cmtNames}{this could be a boolean indicating that the
compartment names should be changed to the named compartments in
the `$MODEL` by `COMP = (name)` (default: `TRUE`). This could
also be a character vector of the compartment names (in order) to
be replaced.}

\item{updateFinal}{Update the parsed model with the model estimates
from the `.lst` output file.}

\item{determineError}{Boolean to try to determine the `nlmixr2`-style residual
error model (like `ipred ~ add(add.sd)`), otherwise endpoints are
not defined in the `rxode2`/`nlmixr2` model (default: `TRUE`)}

\item{validate}{Boolean that this tool will attempt to "validate"
the model by solving the derived model under pred conditions
(etas are zero and eps values are zero)}

\item{lst}{the NONMEM output extension, defaults to `.lst`}

\item{ext}{the NONMEM ext file extension, defaults to `.ext`}
}
\value{
rxode2 function
}
\description{
Convert a NONMEM source file to a rxode model (nlmixr2-syle)
}
\examples{

nonmem2rx(system.file("run001.mod", package="nonmem2rx"))

nonmem2rx(system.file("mods/cpt/runODE032.ctl", package="nonmem2rx"), lst=".res")

}
