% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R, R/records.R, R/theta.R
\name{nonmem2rxRec.ome}
\alias{nonmem2rxRec.ome}
\alias{nonmem2rxRec}
\alias{nonmem2rxRec.default}
\alias{nonmem2rxRec.the}
\title{Record handling for nonmem records}
\usage{
\method{nonmem2rxRec}{ome}(x)

nonmem2rxRec(x)

\method{nonmem2rxRec}{default}(x)

\method{nonmem2rxRec}{the}(x)
}
\arguments{
\item{x}{Nonmem record data item, should be of class c(stdRec,
"nonmem2rx") where the stdRec is the standardized record (pro for
`$PROBLEM`, etc)}
}
\value{
Nothing, called for side effects
}
\description{
Record handling for nonmem records
}
\details{
Can add record parsing and handling by creating a S3 method for each type of standardized method
}
\author{
Matthew L. Fidler
}
\keyword{internal}
