% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rxode2}
\alias{rxode}
\alias{RxODE}
\alias{\%>\%}
\alias{ini}
\alias{model}
\alias{rxRename}
\alias{rxSolve}
\alias{rxUiGet}
\alias{logit}
\alias{expit}
\alias{model<-}
\alias{lotri}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{lotri}{\code{\link[lotri]{lotri}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rxode2}{\code{\link[rxode2:logit]{expit}}, \code{\link[rxode2]{ini}}, \code{\link[rxode2]{logit}}, \code{\link[rxode2]{model}}, \code{\link[rxode2:model-set]{model<-}}, \code{\link[rxode2:rxode2]{rxode}}, \code{\link[rxode2:rxode2]{RxODE}}, \code{\link[rxode2]{rxode2}}, \code{\link[rxode2]{rxRename}}, \code{\link[rxode2]{rxSolve}}, \code{\link[rxode2]{rxUiGet}}}
}}

